feat: implement basic redux store structure
This borrows from the chirper example project to create the basic
project structure with actions, components, middleware, reducers,
and utils directories as the source tree. The main app component gets
initial data from the fake server, and dispatches actions for users and
questions. For now, it just logs the redux store to the console upon
updating the users and questions state.