feat: add login page component
This component is a page with some markup to display a welcome greeting,
image, and login component.

The login component now has a button to perform the setAuthedUser
action, instead of changing the authed user whenever the select option
changes.

After logging in, now present an authenticated app component, with a
logout button that clears the authenticated user and returns to the
login page.